+++ example/index.ts
@@ -1,4 +1,6 @@
 import { Elysia, t } from 'elysia'
+import z from 'zod'
+
 import { openapi, withHeaders } from '../src/index'
 
 const schema = t.Object({
@@ -74,11 +76,65 @@ export const app = new Elysia()
 		}),
 		{
 			parse: ['json', 'formdata'],
-			body: 'user',
+			body: 'schema',
 			response: {
 				200: 'schema',
-				400: 'schema2'
+				400: z.object({
+					a: z.string(),
+					b: z.literal('a')
+				})
 			}
+		}
+	)
+	.post(
+		'/json/:id',
+		({ body, params: { id }, query: { name, email, birthday } }) => ({
+			...body,
+			id,
+			name,
+			email,
+			birthday
+		}),
+		{
+			params: t.Object({
+				id: t.String()
+			}),
+			query: t.Object({
+				name: t.String(),
+				email: t.String({
+					description: 'sample email description',
+					format: 'email'
+				}),
+				birthday: t.String({
+					description: 'sample birthday description',
+					pattern: '\\d{4}-\\d{2}-\\d{2}',
+					minLength: 10,
+					maxLength: 10
+				})
+			}),
+			body: t.Object({
+				username: t.String(),
+				password: t.String()
+			}),
+			response: t.Object(
+				{
+					username: t.String(),
+					password: t.String(),
+					id: t.String(),
+					name: t.String(),
+					email: t.String({
+						description: 'sample email description',
+						format: 'email'
+					}),
+					birthday: t.String({
+						description: 'sample birthday description',
+						pattern: '\\d{4}-\\d{2}-\\d{2}',
+						minLength: 10,
+						maxLength: 10
+					})
+				},
+				{ description: 'sample description 3' }
+			)
 		}
 	)
 	.get('/id/:id/name/:name', () => {})